@@ -5,7 +5,7 @@ homepage: https://romatroskin.github.io/
 repository: https://github.com/romatroskin/social_share_plugin.git
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.0 <2.0.0"
 
 dependencies:
